requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
lxml==4.9.3
python-dotenv==1.0.0
gunicorn==21.2.0